refactor(runtime): extract SwapState into its own module
Pure refactor: no behaviour change. Moves the single-swap strategy's state
and its ~150-line startup driver out of manager.rs into a dedicated
process/swap_state.rs, taking manager.rs from 864 lines to 300 and leaving
it doing dispatch only. The state and the launch sequence that mutates it
were previously separated by several hundred lines of unrelated code.

Collapses slot_save_path, cache_reuse, cache_type_k and cache_type_v into a
single standing ServerConfigOptions template. Every launch now resolves to

    template  ⊕  per-call overrides  ⊕  this request's context chain

which drops new_single_swap from 8 arguments to 5 and removes its
too_many_arguments allow. The context chain is assigned rather than
overlaid, since the manager is authoritative for all three rungs and an
inherited model_server_ctx would silently size a launch for another model.

The template is the point: a flag added to ServerConfigOptions now reaches
llama-server through this path with no change here at all. Previously each
flag needed its own strategy field, constructor parameter and driver line —
which is why --mlock never reached the proxy launch path in the first place,
and why `gglib serve` grew a parallel command builder rather than a fifth
parameter.

ProcessStrategy::SingleSwap is boxed so the options template does not
inflate every ProcessManager, including the Concurrent case that has no use
for it.
